Enable SAL for platform component on Windows User Mode (#1306)
* Fix SAL mismatches in OpenSSL and platform specific code

Since in user mode we don't build with SAL tracked, many of these are missed

* Enable SAL in Platform, fix review comments

* Fix casting

* Add assert that event creation doesn't fail, fix a few more annotation things

* Do assert in debug only

* Clean up do whiles

* 1 more fix

* Add wrapper to CloseHandle to avoid SAL bug

* Switch everything to CxPlatCloseHandle, move to header

+++ src/platform/tls_schannel.c
@@ -1963,7 +1963,7 @@ CxPlatTlsWriteDataToSchannel(
                 SchannelKeyReady,
                 TlsContext->Connection,
                 "Key Ready Type, %u [%hu to %hu]",
-                TrafficSecret->TrafficSecretType,
+                (uint32_t)TrafficSecret->TrafficSecretType,
                 TrafficSecret->MsgSequenceStart,
                 TrafficSecret->MsgSequenceEnd);
             if (TlsContext->IsServer) {
